Implemented changes in Joi validation separating the schemas the schema with the validator
